Refactor code and adjust UI settings
The commit mainly makes two types of changes. On one hand, it simplifies a mathematical calculation in the 'StageControl.java' file to make the code cleaner and easier to read. On the other hand, it modifies certain UI settings in the 'Main.fxml' file, making the 'tables_hbox' always grow in 'VBox' and modifying the way rotation of image views are handled in the 'TableController.java' file. It also removes an unused variable 'hide_im'.

+++ src/main/java/com/girbola/StageControl.java
@@ -64,7 +64,7 @@ public class StageControl extends Stage {
                 double asitisWidth = 0;
                 double hiddenWidths = ((3 - visibles) * 50);
                 double tablesRootPane = (modelMain.tables().getTables_rootPane().getPrefWidth()-145);
-                double divider = Math.floor(((double) tablesRootPane) / (double) visibles);
+                double divider = Math.floor(tablesRootPane / (double) visibles);
                 divider -= hiddenWidths;
 
                 //divider -= (hiddenWidths*3);
